Email_Checker_Bool, Email_Checker_List and Email_Piruk modules
the Email_Checker_Bool Function gets a string and returns true if it contains a "valid" email address (it only checks the prefix and @ and not the domain) , the function returns False if the string doesn't have an email address in it.

the Email_Checker_List Function gets a string and returns a list of all VALID email addressed (again, by prefix and @ only).

The Email_Piruk function receives a string, gets a list of all VALID email addresses in it (by using the Email_Checker_List function) and disassembles each valid email to its prefix, domain (everything  from the "@" to the first "."), and everything after the domain.
These 3 sub_strings are then put into a list and that list is put into a bigger list which eventually contains all disassembles valid emails, which the function then returns